requirements.txt
clickhouse-connect
pandas
python-dotenv
requests